pyproject.toml
[project]
name = "grivredr"
version = "0.1.0"
description = "AI-powered web scraper generator for government grievance portals"
requires-python = ">=3.11"
dependencies = [
    # Browser Automation
    "browser-use>=0.11.0",
    "playwright>=1.49.0",

    # AI/ML
    "anthropic>=0.39.0",
    "openai>=1.12.0",

    # Web Scraping & Parsing
    "beautifulsoup4>=4.12.3",
    "lxml>=5.2.0",

    # Data & Config
    "python-dotenv>=1.0.1",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",

    # Async & IO
    "aiofiles>=23.2.1",
    "httpx>=0.26.0",

    # Utilities
    "pillow>=10.3.0",
    "loguru>=0.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "black>=24.1.1",
    "flake8>=7.0.0",
]

web = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "flask>=3.0.0",
    "flask-socketio>=5.3.6",
    "python-socketio>=5.11.0",
    "jinja2>=3.1.3",
    "python-multipart>=0.0.6",
]

langchain = [
    "langchain>=0.3.14",
    "langchain-anthropic>=0.3.3",
    "langchain-community>=0.3.14",
]

vectordb = [
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
]

all = [
    "grivredr[dev,web,langchain,vectordb]",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "black>=24.1.1",
    "flake8>=7.0.0",
]
